requirements.txt
discord.py>=2.3.2
pymysql>=1.1.0
python-dateutil>=2.8.2
pytz>=2024.1
PyNaCl>=1.5.0
recommonmark>=0.7.1